dice are working!
