+++ public/css/style.css
@@ -136,20 +136,26 @@ button:hover {
   color: white;
 }
 
-#dice-button {
-  width: 45px;
-  height: 45px;
-  padding: 0%;
+#dice-buttons {
+  width: 40px;
+  height: 40px;
+  padding: 0.05%;
   border-radius: 10px;
   border: 2px solid red;
 }
 
-#dice-button:hover {
+#dice-buttons:hover {
   transform: scale(1.06);
 }
 
 #d20-image {
-  width: 100%;
-  height: 100%;
+  width: 35px;
+  height: 35px;
   border-radius: 10px;
+}
+
+.dice-text {
+  font-size: 20px;
+  color: deepskyblue;
+  background-color: darkolivegreen;
 }
