+++ seeds/characterData.json
@@ -10,7 +10,7 @@
     "user_id": 1
   },
   {
-    "id": 7,
+    "id": 9,
     "name": "b",
     "type": "warrior",
     "position": {
